requirements.txt
pytest==7.2.0
pytest-rerunfailures    #用例失败后重跑
allure-pytest  #用于生成美观的测试报告
pyyaml
selenium==4.7.2
Appium-Python-Client~=2.2.0